+++ test/aqua/test_graph_partition.py
@@ -19,8 +19,8 @@ import numpy as np
 from qiskit import BasicAer
 from qiskit.aqua import run_algorithm, aqua_globals
 from qiskit.aqua.input import EnergyInput
-from qiskit.aqua.translators.ising import graph_partition
-from qiskit.aqua.translators.ising.common import random_graph, sample_most_likely
+from qiskit.optimization.ising import graph_partition
+from qiskit.optimization.ising.common import random_graph, sample_most_likely
 from qiskit.aqua.algorithms import ExactEigensolver
 
 
@@ -32,7 +32,7 @@ class TestGraphPartition(QiskitAquaTestCase):
         aqua_globals.random_seed = 100
         self.num_nodes = 4
         self.w = random_graph(self.num_nodes, edge_prob=0.8, weight_range=10)
-        self.qubit_op, self.offset = graph_partition.get_qubit_op(self.w)
+        self.qubit_op, self.offset = graph_partition.get_operator(self.w)
         self.algo_input = EnergyInput(self.qubit_op)
 
     def _brute_force(self):