+++ test/aqua/test_partition.py
@@ -19,8 +19,8 @@ import numpy as np
 from qiskit import BasicAer
 from qiskit.aqua import run_algorithm
 from qiskit.aqua.input import EnergyInput
-from qiskit.aqua.translators.ising import partition
-from qiskit.aqua.translators.ising.common import read_numbers_from_file, sample_most_likely
+from qiskit.optimization.ising import partition
+from qiskit.optimization.ising.common import read_numbers_from_file, sample_most_likely
 from qiskit.aqua.algorithms import ExactEigensolver
 
 
@@ -31,7 +31,7 @@ class TestSetPacking(QiskitAquaTestCase):
         super().setUp()
         input_file = self._get_resource_path('sample.partition')
         number_list = read_numbers_from_file(input_file)
-        qubit_op, _ = partition.get_qubit_op(number_list)
+        qubit_op, _ = partition.get_operator(number_list)
         self.algo_input = EnergyInput(qubit_op)
 
     def test_partition(self):