+++ test/chemistry/test_end2end_with_iqpe.py
@@ -26,7 +26,7 @@ from qiskit.aqua.algorithms.classical import ExactEigensolver
 from qiskit.aqua.operators import Z2Symmetries
 from qiskit.chemistry.drivers import PySCFDriver, UnitsType
 from qiskit.chemistry import FermionicOperator, QiskitChemistryError
-from qiskit.chemistry.aqua_extensions.components.initial_states import HartreeFock
+from qiskit.chemistry.components.initial_states import HartreeFock
 
 
 class TestIQPE(QiskitChemistryTestCase):