+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Changelog](http://keepachangelog.com/en/1.0.0/).
 > -   **Fixed**: for any bug fixes.
 > -   **Security**: in case of vulnerabilities.
 
-[UNRELEASED](https://github.com/Qiskit/qiskit-aqua/compare/0.6.0...HEAD)
+[UNRELEASED](https://github.com/Qiskit/qiskit-aqua/compare/0.6.1...HEAD)
 ========================================================================
 
 Changed
@@ -23,6 +23,11 @@ Changed
 
 -   `VQE`, `VQC` and `QSVM` now use parameterized circuits if it is available to save time 
     in transpilation. (#693)
+-   Refactor of ai,finance and optimization. (#700) Among other changes:
+    - qiskit/aqua/translators/data_providers/ moved to qiskit/finance/data_providers/
+    - qiskit/aqua/translators/ising/portfolio.py and portfolio_diversification.py moved to qiskit/finance/ising/
+    - qiskit/aqua/translators/ising/ (i.e. all but above 2) moved to qiskit/optimization/ising/
+-   UCCSD updated to all excitation pool to be managed by an adaptive algorithm like VQEAdapt. (#685)
 
 Added
 -----
@@ -33,7 +38,11 @@ Added
     which a `Grover` object will now check for, upon initialization, on its provided oracle.  (#681)
 -   `VariationalForm` and `FeatureMap` has a new property on `support_parameterized_circuit`, which 
     implies whether or not can be built with `Parameter` (or `ParameterVector`). Furthermore, 
-    the `evolution_instruction` method support `Parameter` as time parameter.  (#693) 
+    the `evolution_instruction` method support `Parameter` as time parameter.  (#693)
+-   VQEAdapt, adaptive version of the VQE algorithm for chemistry. (#685)
+-   Optionally split qobj by max gates per job. (#694)
+-   An option in evolution_instruction to be able to control whether or not add a barrier
+    for every slice. (#708)
 
 Fixed
 -------
@@ -44,6 +53,14 @@ Removed
 -------
 
 -   The `CircuitCache` class is removed, use parameterized circuits as an alternative. (#693)
+
+[0.6.1](https://github.com/Qiskit/qiskit-aqua/compare/0.6.0...0.6.1) - 2019-10-16
+=================================================================================
+
+Changed
+-------
+
+-   Remove terra cap from stable branch. (#709)
 
 [0.6.0](https://github.com/Qiskit/qiskit-aqua/compare/0.5.5...0.6.0) - 2019-08-22
 =================================================================================